+++ hosts/darwin/codex/default.nix
@@ -3,6 +3,7 @@
   nvf,
   catppuccin-bat,
   catppuccin,
+  nix-secrets,
   ...
 }:
 
@@ -150,4 +151,6 @@ in
     ];
     syncthingDeploy = true; # copy renewed cert to syncthing and restart on renewal
   };
+
+  nixie.krb5.keytabFile = "${nix-secrets}/keytab-codex.age";
 }