+++ hosts/nixos/common-nixos.nix
@@ -28,6 +28,7 @@ in
     ../../modules/common/github-secrets.nix
     ../../modules/common/tailscale-secrets.nix
     ../../modules/common/cachix-secrets.nix
+    ../../modules/common/krb5-client.nix
   ];
 
   networking.useDHCP = true;