+++ hosts/nixos/gammu/default.nix
@@ -1,4 +1,4 @@
-{ lib, ... }:
+{ lib, nix-secrets, ... }:
 
 let
   userDefs = import ../../../users.nix;
@@ -79,4 +79,6 @@ in
     ];
     syncthingDeploy = true;
   };
+
+  nixie.krb5.keytabFile = "${nix-secrets}/keytab-gammu.age";
 }